refactor(storage): extract _AuditMixin (insert_audit_row + insert_raw_payload)
