+++ src/uw_scan/storage/repository.py
@@ -37,6 +37,7 @@ from ._helpers import (
     redact_params,
     status_family_for,
 )
+from .audit import _AuditMixin
 
 # Row dataclasses live in rows.py since the PR-1 split. Re-exported here so
 # existing callers (`from uw_scan.storage.repository import JobRow`) continue
@@ -329,7 +330,7 @@ _RECORD_HEALTH_EXCLUDED_TABLES = {
 logger = logging.getLogger(__name__)
 
 
-class Repository(_BaseMixin):
+class Repository(_AuditMixin, _BaseMixin):
     """Repository wraps a psycopg connection and exposes typed CRUD.
 
     Inherits __init__ and the conn property from _BaseMixin. As PR-1/2/3
@@ -396,62 +397,6 @@ class Repository(_BaseMixin):
     def release_advisory_lock(self, key: int) -> None:
         with self._conn.cursor() as cur:
             cur.execute("SELECT pg_advisory_unlock(%s)", (key,))
-
-    # ------------------------------------------------------------------
-    # api_request_audit + raw_payloads
-    # ------------------------------------------------------------------
-    def insert_audit_row(
-        self,
-        run_id: int,
-        endpoint_slug: str,
-        endpoint_path: str,
-        params: dict[str, Any],
-        status_code: int,
-        started_at: datetime,
-        finished_at: datetime,
-        daily_req_count: int | None,
-        minute_req_remaining: int | None,
-        minute_req_reset: str | None,
-        error_message: str | None = None,
-    ) -> int:
-        sql = (
-            f"INSERT INTO {self._schema}.api_request_audit ("
-            "run_id, endpoint_slug, endpoint_path, params_json, status_code, "
-            "request_started_at, request_finished_at, daily_req_count, "
-            "minute_req_remaining, minute_req_reset, error_message) "
-            "VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) RETURNING audit_id"
-        )
-        with self._conn.cursor() as cur:
-            cur.execute(
-                sql,
-                (
-                    run_id,
-                    endpoint_slug,
-                    endpoint_path,
-                    Jsonb(params),
-                    status_code,
-                    started_at,
-                    finished_at,
-                    daily_req_count,
-                    minute_req_remaining,
-                    minute_req_reset,
-                    error_message,
-                ),
-            )
-            row = cur.fetchone()
-        assert row is not None
-        return int(row[0])
-
-    def insert_raw_payload(self, audit_id: int, payload: dict | list) -> int:
-        sql = (
-            f"INSERT INTO {self._schema}.raw_payloads (audit_id, payload_jsonb) "
-            "VALUES (%s, %s) RETURNING payload_id"
-        )
-        with self._conn.cursor() as cur:
-            cur.execute(sql, (audit_id, Jsonb(payload)))
-            row = cur.fetchone()
-        assert row is not None
-        return int(row[0])
 
     # ------------------------------------------------------------------
     # external_api_requests
